Move env patching fixture to conftest
